room db

+++ app/src/main/java/com/orensharon/brainq/data/event/RequestStateChangedEvent.java
@@ -3,11 +3,11 @@ package com.orensharon.brainq.data.event;
 import java.util.Date;
 
 public class RequestStateChangedEvent {
-    public final int requestId;
+    public final long requestId;
     public final long ts;
     public final boolean state;
 
-    public RequestStateChangedEvent(int requestId, long ts, boolean state) {
+    public RequestStateChangedEvent(long requestId, long ts, boolean state) {
         this.requestId = requestId;
         this.ts = ts;
         this.state = state;
